requirements.txt
Django==4.2.2
psycopg2-binary==2.9.6
celery==5.3.0
redis==3.5.3
djangorestframework==3.14.0
django-celery==3.1.17
django-redis==5.2.0
django-redis-cache==3.0.1
django-filter==23.2
django-celery-beat==2.5.0